Removed v4 files
