+++ pom.xml
@@ -15,7 +15,6 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <scala.version>2.11.0</scala.version>
         <jaxrs.version>2.0</jaxrs.version>
 
 
@@ -57,13 +56,6 @@
             <version>2.3</version>
         </dependency>
 
-        <!-- needed to support scala -->
-        <dependency>
-            <groupId>org.joda</groupId>
-            <artifactId>joda-convert</artifactId>
-            <version>1.5</version>
-        </dependency>
-
         <!-- for base64 -->
         <dependency>
             <groupId>commons-codec</groupId>
@@ -88,54 +80,6 @@
             <type>maven-plugin</type>
         </dependency>
 
-        <!-- test dependencies -->
-        <dependency>
-            <groupId>info.cukes</groupId>
-            <artifactId>cucumber-scala_2.10</artifactId>
-            <version>1.1.5</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>info.cukes</groupId>
-            <artifactId>cucumber-junit</artifactId>
-            <version>1.1.5</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.rubiconproject.oss</groupId>
-            <artifactId>jchronic</artifactId>
-            <version>0.2.6</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.scala-lang</groupId>
-            <artifactId>scala-library</artifactId>
-            <version>${scala.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.scala-lang</groupId>
-            <artifactId>scala-compiler</artifactId>
-            <version>${scala.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.scalatest</groupId>
-            <artifactId>scalatest_2.11</artifactId>
-            <version>2.1.3</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.scalatest</groupId>
-            <artifactId>scalatest-maven-plugin</artifactId>
-            <version>1.0</version>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-all</artifactId>
-            <version>1.9.5</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <packaging>jar</packaging>
@@ -165,47 +109,6 @@
                     <source>1.5</source>
                     <target>1.5</target>
                 </configuration>
-            </plugin>
-
-            <plugin>
-                <groupId>org.scala-tools</groupId>
-                <artifactId>maven-scala-plugin</artifactId>
-                <version>2.15.2</version>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>add-source</goal>
-                            <goal>compile</goal>
-                            <goal>testCompile</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-
-            <plugin>
-                <groupId>org.scalatest</groupId>
-                <artifactId>scalatest-maven-plugin</artifactId>
-                <version>1.0</version>
-                <configuration>
-                    <systemProperties>
-                        <net.conjur.integrationTestsEnabled>
-                            ${env.CONJUR_INTEGRATION_TESTS_ENABLED}
-                        </net.conjur.integrationTestsEnabled>
-                        <net.conjur.api.credentials>${env.CONJUR_CREDENTIALS}</net.conjur.api.credentials>
-                        <net.conjur.api.account>${env.CONJUR_ACCOUNT}</net.conjur.api.account>
-                        <net.conjur.api.stack>${env.CONJUR_STACK}</net.conjur.api.stack>
-                        <net.conjur.api.resource.requestLogging>true</net.conjur.api.resource.requestLogging>
-                        <net.conjur.api.authn.requestLogging>true</net.conjur.api.authn.requestLogging>
-                    </systemProperties>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>test</id>
-                        <goals>
-                            <goal>test</goal>
-                        </goals>
-                    </execution>
-                </executions>
             </plugin>
 
             <plugin>
