+++ src/main/java/net/conjur/api/AuthnProvider.java
@@ -1,4 +1,4 @@
-package net.conjur.apiV5;
+package net.conjur.api;
 
 
 /**
