--- src/main/java/net/conjur/api/examples/SecretsManagement.java
@@ -1,7 +1,0 @@
-package net.conjur.api.examples;
-
-/**
- * This example shows how to manage secrets using Conjur variables.  You can run it with mvn exec:net.conjur.api.examples.SecretsManagement
- */
-public class SecretsManagement {
-}
